Fixed all int-to-pointer and pointer-to-int cast warnings by changing int to long datatype. Changed all int variabls to long that represent a pointer. Fixed implicit-int and implicit-function-declaration warnings by adding function prototypes and defining function return type. Changed argument type in error() to void * and added function prototype. Fixed EOF (-1) switch out of range warning by casting #define MYEOF (char)EOF. Fixed decrement operation bug in VARpop macro which caused SEGFAULT.

+++ map.c
@@ -8,7 +8,7 @@ Vmapa ()  /* -------------------------------- (mapa 'fun 'list1 ... 'listn) */
 {
    kerncell fun = ARGnum1;
    int	    arg1 = ARGidx1 + 1;
-   int	    argn = argtop;
+   long	    argn = argtop;
    register int i;
    kerncell res = argstk[arg1];
 
@@ -37,7 +37,7 @@ Vmapcar ()  /* ---------------------------- (mapcar 'fun 'list1 ... 'listn) */
    kerncell fun = ARGnum1;
    kerncell res = NIL;
    int	    arg1 = ARGidx1 + 1;
-   int	    argn = argtop;
+   long	    argn = argtop;
    register int i;
 
 	CHECKvargs1(mapcarsym,2);		/* at least 2 args required */
@@ -64,7 +64,7 @@ Vmapd ()  /* -------------------------------- (mapd 'fun 'list1 ... 'listn) */
 {
    kerncell fun = ARGnum1;
    int	    arg1 = ARGidx1 + 1;
-   int	    argn = argtop;
+   long	    argn = argtop;
    register int i;
    kerncell res = argstk[arg1];
 
@@ -93,7 +93,7 @@ Vmapcdr ()  /* ---------------------------- (mapcdr 'fun 'list1 ... 'listn) */
    kerncell fun = ARGnum1;
    kerncell res = NIL;
    int	    arg1 = ARGidx1 + 1;
-   int	    argn = argtop;
+   long	    argn = argtop;
    register int i;
 
 	CHECKvargs1(mapcdrsym,2);		/* at least 2 args required */
